requirements.txt
aiohttp==3.5.4
uvicorn==0.7.1
fs==2.4.5
APScheduler==3.6.0
gmqtt==0.5.6
easyconf==1.1
appdirs==1.4.3
loguru==0.3.2
nested_lookup==0.2.12
fastapi==0.30.0
typing==3.7.4
PyYAML==5.1.1
fs.smbfs==0.5.2
urllib3==1.24.3
requests==2.21.0
kodi-json==1.0.0
pycryptodome==3.9.4
python-magichue==0.2.7



